requirements.txt
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==5.5
cycler==0.11.0
decorator==4.4.2
deprecation==2.1.0
eyed3==0.9.7
filetype==1.2.0
fonttools==4.38.0
idna==3.6
imageio==2.31.2
imageio-ffmpeg==0.4.9
importlib-metadata==6.7.0
kiwisolver==1.4.5
matplotlib==3.5.3
MouseInfo==0.1.3
moviepy==1.0.3
mutagen==1.47.0
numpy==1.21.6
opencv-python==4.9.0.80
packaging==23.2
Pillow==9.5.0
proglog==0.1.10
PyAutoGUI==0.9.54
pycryptodomex==3.20.0
pydub==0.25.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
pytube==15.0.0
pytweening==1.0.7
requests==2.31.0
rubicon-objc==0.4.6
six==1.16.0
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
websockets==11.0.3
youtube-dl==2021.12.17
yt-dlp==2023.11.16
zipp==3.15.0
